NUM-2072 move error handling in controller in order to not override for all controllers the message for missing request body
