+++ src/main/java/de/vitagroup/num/web/controller/CohortController.java
@@ -9,6 +9,7 @@ import de.vitagroup.num.domain.dto.TemplateSizeRequestDto;
 import de.vitagroup.num.mapper.CohortMapper;
 import de.vitagroup.num.service.CohortService;
 import de.vitagroup.num.service.exception.CustomizedExceptionHandler;
+import de.vitagroup.num.service.exception.dto.ErrorDetails;
 import de.vitagroup.num.service.logger.AuditLog;
 import de.vitagroup.num.web.config.Role;
 import io.swagger.annotations.ApiOperation;
@@ -18,7 +19,11 @@ import java.util.stream.Collectors;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -30,7 +35,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.WebRequest;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 @RequestMapping(value = "/cohort", produces = "application/json")
@@ -127,5 +134,25 @@ public class CohortController extends CustomizedExceptionHandler {
     return ResponseEntity.ok(
         cohortService.getCohortGroupSizeWithDistribution(
             cohortGroupDto, principal.getSubject(), allowUsageOutsideEu));
+  }
+
+  /**
+   * Note : is on controller level to avoid overriding behaviour for all controllers
+   * @param ex
+   * @param headers
+   * @param status
+   * @param request
+   * @return
+   */
+  @Override
+  protected ResponseEntity<Object> handleHttpMessageNotReadable(HttpMessageNotReadableException ex, HttpHeaders headers, HttpStatus status, WebRequest request) {
+    log.warn("Request for {} failed with error message {} ", request.getDescription(false), ex.getMessage());
+    Map<String,String> errors = Map.of("Error", "error.missing_request_body");
+    ErrorDetails errorDetails = ErrorDetails
+            .builder()
+            .message( "Request body is required" )
+            .details( errors )
+            .build();
+    return ResponseEntity.status( HttpStatus.BAD_REQUEST ).body( errorDetails );
   }
 }
